requirements.txt
defusedxml==0.7.1
django-allauth==0.55.0
django-extensions==3.2.3
django-sslserver==0.22
MarkupSafe==2.1.3
python3-openid==3.2.0
requests-oauthlib==1.3.1
Werkzeug==2.3.7
